updated from dagor4 repo on 2023/11/07, rev 734cc7c4ec4235531e627319981a91c135d4319f

+++ prog/gameLibs/publicInclude/render/daBfg/detail/access.h
@@ -11,7 +11,7 @@
 namespace dabfg
 {
 
-enum class Access : uint32_t
+enum class Access : uint8_t
 {
   UNKNOWN,
   READ_ONLY,
